+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Workaround for Solaris SPARC (#2).
 * Remove ignored edges in `clone()` (#6).
+* Add ptional verbose info in `plot.trajectory()` labels (#5).
 
 # simmer.plot 0.1.8